Add unit test for update

+++ cmd/je/update.go
@@ -2,11 +2,12 @@ package main
 
 import (
 	"encoding/json"
+	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 	"github.com/wooos/je/cmd/je/require"
 	"github.com/wooos/je/pkg/parser"
 )
@@ -16,29 +17,35 @@ This command update a json file.
 
 `
 
-var (
+type updateOptions struct {
 	setArrays []string
-)
+	dryRun bool
+}
 
-func newUpdateCmd() *cobra.Command {
+func newUpdateCmd(out io.Writer) *cobra.Command {
+	o := updateOptions{}
 	cmd := &cobra.Command{
 		Use:   "update [FILENAME]",
 		Short: "update a json file",
 		Args:  require.ExactArgs(1),
 		Long:  updateDesc,
-		RunE:  runUpdate,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return o.runUpdate(out, cmd, args)
+		},
 	}
 
-	addUpdateFlags(cmd.Flags())
+	cmd.SetOut(out)
+
+	flags := cmd.Flags()
+	flags.StringSliceVar(&o.setArrays, "set", []string{}, "set values on the command line (can specify multiple or separate values with commas: key1=val1,key2=val2)")
+	flags.BoolVar(&o.dryRun, "dry-run", false, "simulate a update")
+
+	cmd.MarkFlagRequired("set")
 
 	return cmd
 }
 
-func addUpdateFlags(f *pflag.FlagSet) {
-	f.StringSliceVar(&setArrays, "set", []string{}, "set values on the command line (can specify multiple or separate values with commas: key1=val1,key2=val2)")
-}
-
-func runUpdate(cmd *cobra.Command, args []string) error {
+func (o *updateOptions) runUpdate(out io.Writer, cmd *cobra.Command, args []string) error {
 	filename := args[0]
 
 	currentMap := map[string]interface{}{}
@@ -51,7 +58,7 @@ func runUpdate(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	for _, val := range setArrays {
+	for _, val := range o.setArrays {
 		if err := parser.ParseInto(val, currentMap); err != nil {
 			return err
 		}
@@ -60,6 +67,11 @@ func runUpdate(cmd *cobra.Command, args []string) error {
 	data, err := json.MarshalIndent(currentMap, "", "  ")
 	if err != nil {
 		return err
+	}
+
+	if o.dryRun {
+		fmt.Fprintln(out, string(data))
+		return nil
 	}
 
 	if err := ioutil.WriteFile(filename, data, os.ModePerm); err != nil {
